refactoring done basically everywhere
